fix(dao): correction de la methode findLast, la methode getInt etait utilisé a la place de getDouble pour les valeurs et le resultat

+++ src/historique/CalculDAOImp.java
@@ -61,10 +61,10 @@ public class CalculDAOImp implements CalculDAO{
             if (rs.next()) {
                 calcul = new Calcul(
                         rs.getInt("id"),
-                        rs.getInt("valeur1Db"),
+                        rs.getDouble("valeur1Db"),
                         rs.getString("operateurDb"),
-                        rs.getInt("valeur2Db"),
-                        rs.getInt("resultatDb")
+                        rs.getDouble("valeur2Db"),
+                        rs.getDouble("resultatDb")
                 );
             }
